The trash drawer rises from the edge instead of appearing
It had no transition at all: the panel and its scrim were mounted through a
portal and unmounted again, so both simply existed or did not.

A bottom sheet that fades in reads as a dialog appearing ABOUT the trash. One
that comes up from the edge it is anchored to reads as a drawer being pulled
out, which is what it is. So the panel carries its own
`view-transition-name` and gets a named animation, while the SCRIM stays
unnamed and cross-fades on the root transition — for a full-bleed wash that
is exactly right, and it is what the trim modal already leans on.

Only one side exists at a time, because the drawer unmounts when closed:
opening has a `new` and no `old`, closing the reverse. Written as two one-way
animations rather than a group tween for that reason. Out is quicker than in
(180ms against 260ms) — arriving wants to be watched, leaving wants to be out
of the way, and an exit as long as its entrance feels like the app holding on
to something you dismissed.

`withViewTransition` MOVED to lib/view-transition.ts rather than being copied.
It had been inline in the trim modal; two copies would be two chances to
forget the root flag it sets, and the e2e's `settleViewTransition` depends on
every caller setting it — a transition that does not announce itself is one
the suite cannot wait for, and that failure lands on a later step as a click
that did nothing.

Both the tile and `onClose` go through it, so the drawer's own X, Escape and
the scrim all exit the same way as the press that opened it.

The module carries a note about where NOT to use this: a change to a
container the rest of the layout is measured against. A view transition
snapshots the whole page, so animating the icon rail's width painted the old
and new layouts on top of each other (#423). Both callers here are overlays.

Verified in the browser: opening animates `::view-transition-new(trash-drawer)`
with no `old`, closing the reverse, the root flag is set for both, and the
panel still clears the rail exactly (left edge 72 = rail width, from the CSS
variable).

app 1227 passed, storybook 242, tsc clean, lint 0 errors.
